Move categories presentation classes to "screens"
